# api-gateway/src/main/resources/static/openapi.yml
openapi: 3.0.3
info:
  title: Taska API Gateway Contract
  version: 1.0.0

servers:
  - url: /
    description: Current host

paths:
  /api/v1/auth/login:
    post:
      summary: Аутентификация пользователя
      operationId: login
      tags:
        - auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequestDto'
      responses:
        '200':
          description: Успешный вход
          headers:
            X-Request-Id:
              $ref: '#/components/headers/X-RequestId-Header'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponseDto'
        default:
          $ref: '#/components/responses/ErrorResponse'

  /api/v1/auth/refresh:
    post:
      summary: Обновление токенов (access/refresh)
      operationId: refresh
      tags:
        - auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequestDto'
      responses:
        '200':
          description: Токены успешно обновлены
          headers:
            X-Request-Id:
              $ref: '#/components/headers/X-RequestId-Header'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefreshResponseDto'
        default:
          $ref: '#/components/responses/ErrorResponse'

  /api/v1/auth/invitations/accept:
    post:
      summary: Принятие инвайта и установка пароля
      operationId: setPasswordByToken
      tags:
        - auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordByTokenRequestDto'
      responses:
        '204':
          description: Пароль успешно изменен (No Content)
          headers:
            X-Request-Id:
              $ref: '#/components/headers/X-RequestId-Header'
        default:
          $ref: '#/components/responses/ErrorResponse'

  /api/v1/users/me:
    get:
      summary: Получение данных текущего пользователя
      operationId: getMyInfo
      tags:
        - auth
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Данные успешно получены
          headers:
            X-Request-Id:
              $ref: '#/components/headers/X-RequestId-Header'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidateAccessTokenResponseDto'
        default:
          $ref: '#/components/responses/ErrorResponse'

  /api/v1/projects:
    post:
      summary: Создание проекта
      operationId: createProject
      tags:
        - project
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectRequestDto'
      responses:
        '201':
          description: Проект успешно создан
          headers:
            X-Request-Id:
              $ref: '#/components/headers/X-RequestId-Header'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponseDto'
        default:
          $ref: '#/components/responses/ErrorResponse'

    get:
      summary: Возвращает проекты текущего пользователя
      operationId: listMyProjects
      tags:
        - project
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Список проектов успешно получен
          headers:
            X-Request-Id:
              $ref: '#/components/headers/X-RequestId-Header'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListMyProjectResponseDto'
        default:
          $ref: '#/components/responses/ErrorResponse'

  /api/v1/projects/{projectId}:
    get:
      summary: Возвращает проект по идентификатору
      operationId: getProject
      tags:
        - project
      security:
        - bearerAuth: []
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Проект успешно найден
          headers:
            X-Request-Id:
              $ref: '#/components/headers/X-RequestId-Header'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponseDto'
        default:
          $ref: '#/components/responses/ErrorResponse'

  /api/v1/projects/{projectId}/members:
    post:
      summary: Добавляет участника в проект
      operationId: addProjectMember
      tags:
        - project
      security:
        - bearerAuth: []
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddProjectMemberRequestDto'
      responses:
        '201':
          description: Участник успешно добавлен
          headers:
            X-Request-Id:
              $ref: '#/components/headers/X-RequestId-Header'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectMemberResponseDto'
        default:
          $ref: '#/components/responses/ErrorResponse'

  /api/v1/projects/{projectId}/members/{userId}:
    patch:
      summary: Меняет роль участника
      operationId: changeProjectMemberRole
      tags:
        - project
      security:
        - bearerAuth: []
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangeProjectMemberRoleRequestDto'
      responses:
        '200':
          description: Роль участника успешно изменена
          headers:
            X-Request-Id:
              $ref: '#/components/headers/X-RequestId-Header'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectMemberResponseDto'
        default:
          $ref: '#/components/responses/ErrorResponse'

    delete:
      summary: Удаляет участника из проекта
      operationId: removeProjectMember
      tags:
        - project
      security:
        - bearerAuth: [ ]
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Участник успешно удален
          headers:
            X-Request-Id:
              $ref: '#/components/headers/X-RequestId-Header'
        default:
          $ref: '#/components/responses/ErrorResponse'

  /api/v1/projects/{projectId}/issues:
    post:
      summary: Создание задачи
      operationId: createIssue
      tags:
        - issue
      security:
        - bearerAuth: [ ]

      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string

        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string

      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateIssueRequestDto'

      responses:
        '201':
          description: Создана новая задача
          headers:
            X-Request-Id:
              $ref: '#/components/headers/X-RequestId-Header'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssueResponseDto'
        default:
          $ref: '#/components/responses/ErrorResponse'

    get:
      summary: Возвращает задачи проекта с фильтрами и пагинацией
      operationId: listIssues
      tags:
        - issue
      security:
        - bearerAuth: [ ]

      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string

        - name: status
          in: query
          required: false
          schema:
            type: string

        - name: assigneeId
          in: query
          required: false
          schema:
            type: string

        - name: page
          in: query
          required: false
          schema:
            type: integer

        - name: pageSize
          in: query
          required: false
          schema:
            type: integer

      responses:
        '200':
          description: Список задач успешно найден
          headers:
            X-Request-Id:
              $ref: '#/components/headers/X-RequestId-Header'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListIssuesResponseDto'
        default:
          $ref: '#/components/responses/ErrorResponse'

  /api/v1/issues/{issueId}:
    get:
      summary: Возвращает задачу с историей
      operationId: getIssue
      tags:
        - issue
      security:
        - bearerAuth: [ ]

      parameters:
        - name: issueId
          in: path
          required: true
          schema:
            type: string

      responses:
        '200':
          description: Задача с историей успешно найдена
          headers:
            X-Request-Id:
              $ref: '#/components/headers/X-RequestId-Header'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssueWithHistoryResponseDto'
        default:
          $ref: '#/components/responses/ErrorResponse'

    put:
      summary: Обновляет основные поля задачи
      operationId: updateIssue
      tags:
        - issue
      security:
        - bearerAuth: [ ]

      parameters:
        - name: issueId
          in: path
          required: true
          schema:
            type: string

      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateIssueRequestDto'

      responses:
        '200':
          description: Данные задачи успешно обновлены
          headers:
            X-Request-Id:
              $ref: '#/components/headers/X-RequestId-Header'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateIssueResponseDto'
        default:
          $ref: '#/components/responses/ErrorResponse'

    delete:
      summary: Выполняет мягкое удаление задачи
      operationId: deleteIssue
      tags:
        - issue
      security:
        - bearerAuth: [ ]

      parameters:
        - name: issueId
          in: path
          required: true
          schema:
            type: string

      responses:
        '204':
          description: Задача успешно удалена (No Content)
          headers:
            X-Request-Id:
              $ref: '#/components/headers/X-RequestId-Header'
        default:
          $ref: '#/components/responses/ErrorResponse'

  /api/v1/issues/{issueId}/assignee:
    put:
      summary: Назначает исполнителя задачи
      operationId: assignIssue
      tags:
        - issue
      security:
        - bearerAuth: [ ]

      parameters:
        - name: issueId
          in: path
          required: true
          schema:
            type: string

      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignIssueRequestDto'

      responses:
        '200':
          description: Исполнитель задачи успешно назначен
          headers:
            X-Request-Id:
              $ref: '#/components/headers/X-RequestId-Header'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssueResponseDto'
        default:
          $ref: '#/components/responses/ErrorResponse'

  /api/v1/issues/{issueId}/transition/{transitionId}:
    put:
      summary: Выполняет переход задачи по workflow (меняет статус задачи)
      operationId: transitionIssue
      tags:
        - issue
      security:
        - bearerAuth: [ ]

      parameters:
        - name: issueId
          in: path
          required: true
          schema:
            type: string

        - name: transitionId
          in: path
          required: true
          schema:
            type: string

      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransitionIssueRequestDto'

      responses:
        '200':
          description: Статус задачи успешно изменен
          headers:
            X-Request-Id:
              $ref: '#/components/headers/X-RequestId-Header'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssueWithHistoryResponseDto'
        default:
          $ref: '#/components/responses/ErrorResponse'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  headers:
    X-RequestId-Header:
      schema:
        type: string
      description: Возвращаемый клиенту сквозной ID запроса

  responses:
    ErrorResponse:
      description: Единая схема ошибки гейтвея
      headers:
        X-Request-Id:
          $ref: '#/components/headers/X-RequestId-Header'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/RestErrorResponse'

  schemas:
    LoginRequestDto:
      type: object
      required: [ email, password ]
      properties:
        email:
          type: string
        password:
          type: string

    LoginResponseDto:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        expiresIn:
          type: integer
          format: int64

    RefreshRequestDto:
      type: object
      required: [ refreshToken ]
      properties:
        refreshToken:
          type: string

    RefreshResponseDto:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        expiresIn:
          type: integer
          format: int64

    PasswordByTokenRequestDto:
      type: object
      required: [ token, newPassword ]
      properties:
        token:
          type: string
        newPassword:
          type: string

    ValidateAccessTokenResponseDto:
      type: object
      properties:
        id:
          type: string
        login:
          type: string
        email:
          type: string
        displayName:
          type: string
        status:
          type: string

    CreateIssueRequestDto:
      type: object
      required: [ issueType, summary, description, priority ]
      properties:
        issueType:
          type: string
        summary:
          type: string
        description:
          type: string
        priority:
          type: string

    IssueResponseDto:
      type: object
      properties:
        id:
          type: string
        projectId:
          type: string
        issueNumber:
          type: integer
        issueKey:
          type: string
        issueType:
          type: string
        summary:
          type: string
        description:
          type: string
        status:
          type: string
        priority:
          type: string
        assigneeId:
          type: string
        reporterId:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        version:
          type: integer

    IssueHistoryResponseDto:
      type: object
      properties:
        id:
          type: string
        eventType:
          type: string
        actorUserId:
          type: string
        occurredAt:
          type: string
          format: date-time
        payload:
          type: object

    IssueWithHistoryResponseDto:
      type: object
      properties:
        issue:
          $ref: '#/components/schemas/IssueResponseDto'
        history:
          type: array
          items:
            $ref: '#/components/schemas/IssueHistoryResponseDto'

    IssueShortResponseDto:
      type: object
      properties:
        id:
          type: string
        issueKey:
          type: string
        summary:
          type: string
        issueType:
          type: string
        priority:
          type: string
        assigneeId:
          type: string

    ListIssuesResponseDto:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/IssueShortResponseDto'
        totalCount:
          type: integer

    UpdateIssueRequestDto:
      type: object
      required: [summary, description, priority]
      properties:
        summary:
          type: string
        description:
          type: string
        priority:
          type: string

    UpdateIssueResponseDto:
      type: object
      properties:
        id:
          type: string
        summary:
          type: string
        description:
          type: string
        priority:
          type: string

    AssignIssueRequestDto:
      type: object
      required: [assigneeId]
      properties:
        assigneeId:
          type: string

    TransitionIssueRequestDto:
      type: object
      properties:
        payload:
          type: object

    CreateProjectRequestDto:
      type: object
      required: [projectKey,name]
      properties:
        projectKey:
          type: string
        name:
          type: string

    ProjectResponseDto:
      type: object
      properties:
        id:
          type: string
        projectKey:
          type: string
        name:
          type: string
        createdBy:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        archivedAt:
          type: string
          format: date-time
          nullable: true

    ListMyProjectResponseDto:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ProjectResponseDto'

    AddProjectMemberRequestDto:
      type: object
      required: [userId,role]
      properties:
        userId:
          type: string
        role:
          type: string
          enum: [ADMIN, MEMBER, VIEWER]

    ChangeProjectMemberRoleRequestDto:
      type: object
      required: [role]
      properties:
        role:
          type: string
          enum: [ ADMIN, MEMBER, VIEWER ]

    ProjectMemberResponseDto:
      type: object
      properties:
        projectId:
          type: string
        userId:
          type: string
        role:
          type: string

    RestErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string